pyproject.toml
[tool.poetry]
name = "toastmasters"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
Django = "^4.1"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
dj-database-url = "^1.0.0"
whitenoise = "^6.2.0"
django-nested-admin = "^3.4.0"
python-dateutil = "^2.8.2"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.24.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pylint = "^2.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
